LPS-59676 updates *DisplayContext.java directory paths in core
