+++ modules/apps/export-import/export-import-web/src/META-INF/resources/init.jsp
@@ -141,7 +141,7 @@ page import="com.liferay.portlet.exportimport.model.ExportImportConfiguration" %
 page import="com.liferay.portlet.exportimport.service.ExportImportConfigurationLocalServiceUtil" %><%@
 page import="com.liferay.portlet.exportimport.staging.LayoutStagingUtil" %><%@
 page import="com.liferay.portlet.exportimport.staging.StagingUtil" %><%@
-page import="com.liferay.portlet.layoutsadmin.context.GroupDisplayContextHelper" %><%@
+page import="com.liferay.portlet.layoutsadmin.display.context.GroupDisplayContextHelper" %><%@
 page import="com.liferay.portlet.trash.util.TrashUtil" %><%@
 page import="com.liferay.taglib.search.ResultRow" %><%@
 page import="com.liferay.taglib.search.SearchEntry" %><%@
