+++ portal-impl/src/com/liferay/portal/comment/display/context/CommentDisplayContextProvider.java
@@ -12,12 +12,13 @@
  * details.
  */
 
-package com.liferay.portal.kernel.comment.context;
+package com.liferay.portal.comment.display.context;
 
 import com.liferay.portal.kernel.comment.Discussion;
 import com.liferay.portal.kernel.comment.DiscussionComment;
 import com.liferay.portal.kernel.comment.DiscussionPermission;
-import com.liferay.portal.kernel.display.context.DisplayContextFactory;
+import com.liferay.portal.kernel.comment.display.context.CommentSectionDisplayContext;
+import com.liferay.portal.kernel.comment.display.context.CommentTreeDisplayContext;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -25,15 +26,13 @@ import javax.servlet.http.HttpServletResponse;
 /**
  * @author Adolfo Pérez
  */
-public interface CommentDisplayContextFactory extends DisplayContextFactory {
+public interface CommentDisplayContextProvider {
 
 	public CommentSectionDisplayContext getCommentSectionDisplayContext(
-		CommentSectionDisplayContext parentCommentSectionDisplayContext,
 		HttpServletRequest request, HttpServletResponse response,
 		DiscussionPermission discussionPermission, Discussion discussion);
 
 	public CommentTreeDisplayContext getCommentTreeDisplayContext(
-		CommentTreeDisplayContext parentCommentTreeDisplayContext,
 		HttpServletRequest request, HttpServletResponse response,
 		DiscussionPermission discussionPermission,
 		DiscussionComment discussionComment);
