+++ portal-impl/src/com/liferay/portal/comment/display/context/DefaultCommentSectionDisplayContext.java
@@ -12,14 +12,14 @@
  * details.
  */
 
-package com.liferay.portal.comment.context;
+package com.liferay.portal.comment.display.context;
 
-import com.liferay.portal.comment.context.util.DiscussionRequestHelper;
-import com.liferay.portal.comment.context.util.DiscussionTaglibHelper;
+import com.liferay.portal.comment.display.context.util.DiscussionRequestHelper;
+import com.liferay.portal.comment.display.context.util.DiscussionTaglibHelper;
 import com.liferay.portal.kernel.comment.Discussion;
 import com.liferay.portal.kernel.comment.DiscussionComment;
 import com.liferay.portal.kernel.comment.DiscussionPermission;
-import com.liferay.portal.kernel.comment.context.CommentSectionDisplayContext;
+import com.liferay.portal.kernel.comment.display.context.CommentSectionDisplayContext;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.theme.ThemeDisplay;
