+++ portal-impl/src/com/liferay/portal/comment/display/context/DefaultCommentTreeDisplayContext.java
@@ -12,15 +12,15 @@
  * details.
  */
 
-package com.liferay.portal.comment.context;
+package com.liferay.portal.comment.display.context;
 
-import com.liferay.portal.comment.context.util.DiscussionRequestHelper;
-import com.liferay.portal.comment.context.util.DiscussionTaglibHelper;
+import com.liferay.portal.comment.display.context.util.DiscussionRequestHelper;
+import com.liferay.portal.comment.display.context.util.DiscussionTaglibHelper;
 import com.liferay.portal.kernel.comment.CommentConstants;
 import com.liferay.portal.kernel.comment.DiscussionComment;
 import com.liferay.portal.kernel.comment.DiscussionPermission;
 import com.liferay.portal.kernel.comment.WorkflowableComment;
-import com.liferay.portal.kernel.comment.context.CommentTreeDisplayContext;
+import com.liferay.portal.kernel.comment.display.context.CommentTreeDisplayContext;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.workflow.WorkflowConstants;
